src/python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepface>=0.0.93",
    "fastapi>=0.116.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "tf-keras>=2.19.0",
    "uvicorn>=0.35.0",
]
